--- app/kumactl/cmd/export/testdata/export-kube.golden.yaml
@@ -1,18 +1,0 @@
-apiVersion: kuma.io/v1alpha1
-kind: Mesh
-metadata:
-  annotations:
-    k8s.kuma.io/mesh-defaults-generated: "true"
-  creationTimestamp: "2024-01-08T17:25:45Z"
-  name: default
-spec: {}
----
-apiVersion: v1
-data:
-  value: LS0tLS1CRUdJTiBSU0EgUFJJVkFURSBLRVktLS0tLQpNSUlFb2dJQkFBS0NBUUVBdGNYWFdySUpjVVpxLzkyOUFMeUNFRWRaM1ZacDQreGVTWlFiMkRFNDEyK1VLMTV1CllUbFNXZDIvZExlSHJMVzQ0b2dUbjNhaUlEQTJiSHBmZUhNNDZtZ3cxWFM4V2JQMlR3UlA4eHVCWmdNTVY4RUYKdUxXQ1UwdTdNYXBNUTBFY05ySU9jeStHYWxkSGlBQTRsN1JkbkNWWnVDaFJsTHhUeDNkcGdiVDEyZGJuWWkwMwp4cmUvQ3MxOTFHZ2FYWWt3ZHI3dFZUdmx6QlVZQkk1U3BQSjlMR0pFVGVUTmNtTkhSS3dWVXpjZGhHdTB4djhICjZpeG05Si9PQ0xUdlZITDlYVlZKZGJXWEFnbTV4NWh4M1pHcHBjdFl2WkFXVWhYeWNNZEk2SmxJamhRSUtZVFEKSlRBcGRFK21RQTIwZER6ZUxiMDZNcTVqNHlVTjV4SE1xVE0yWlFJREFRQUJBb0lCQUdROWh5WkgyamN5ZXgxQQovZEtqSWlLeUJsZXVLTlQ5YTMydUJMRHpiQUk2VVA2ZG84ZFA0aDhXejFjaFF0dTJsdyt0dFF5U1VwUG5vb3V0CjdHb3pPR2JrTUZIMmYwd0VsMGM1L0pEM1VTT2RlTGdJQlF5NFkvZGxZd1RmZ2hEd25ndDNvOHBYb2d1UXVDekMKMm5LbWo1emxhQlhMZmpuRXBYcnhXaWlpVDVYaUk0UGxNY2lBQUJPbDlxT1BJRjNKV2toRXhEM2pwMDZTcDdnNQpzaWlUTXNZMVhVSkljZlRtWVREQW5QVDVBSk9LdmtPaGNyVkNhWjl3QnZKK1E1ZlJta2plcGtjUTE5MWVlM3NhCm1jZzNGZXJaVnlqbDlFd0FSUDRCUDlvejAxNkFPWXdCTEdyRnhoTm51WDltUUpIOUdocm5HeEV0QTczZkdtd24KNUlqZW56MENnWUVBNjBnSTRsMW10b0F4NGJha2JDbTY0MDJnME1xbHZlamg0QUVhTExWN2xiZWUxSS8yVm0wbAp0aUJSSzUvSm45Zm85d0N0NzFoYUljaURiQkw4VmdwRGRJQTAxWHMwRXN3M05RYnFuSHRKZEdwZGdzYXArRGZQCkFmYUk0ZC9wdFBPZ0pOMXFEb2tyTmdkSm50cUNTcVhLVXJGUFNUUk02Tkw1TUx0Mlh2RWJyU2NDZ1lFQXhjZVAKR29CTnVLb3ZhcHRlU2loUmtlU3A0a3daQ25mNnVoUkNkS1RVNGo4NEliVDY3K0dzS29ESmtrSjRNbk9rcHBoaQpGL29IOFJSaVVqUG5pL2NqamxDTEs0WllUSCtBUms4Qks0KzhxaWFTNDRxa21pV2JRb1BnSEJUa2pJMlZ5QTJUCkZCYXNjQzFJSDk1cXZPbjBXTS9GVitDcVhMREVQVUVmeW1SNmo1TUNnWUFRNkw1cXRsZTVhZ0VydlpxWGczQkIKQWNVNi9EVnBMcUY5SGRjcHUybEVXMHY0N3RlL1lsekJlQlVkVTRDRUVscFYxaUNxQU9tSllBWEI4S05kODZMaApibzFkZ2xzOXRwdUpIaFFpWnVIaWtmcGllYkNoZkRGT2JNOXRuRkFKaHpETk5YUjMwdmZlREN3SVZKL2tpWUcrCjY2dE5FWXN0ZUd3Y1NueTVXZzZwSlFLQmdCUkNPYlZVT1NIV1JOOGRwUTNoaldGdFU3MTdZQUZEZ3VvMmt3d1IKcWtDMml4QmcwbWRwN1ZHNU9XeUM1SDcrZTNvWkY4c3U4RXRRZERYSlhVbG4ranRrb3F6RnFSdkUzNlhVQitHbgphb0xmU3FLR0hCQXlFL0kvUWVCK2YxNnFUaEpjMlpqd0xaK2I4OHhYNHdVUXpkZWZDSzdVczM2bEdLVFJtK2tHClc4dURBb0dBWVZnOWp5a3IrNlpYMkU1V0Y1MHl4NTNERGsxNk54elFSVEhVYWlxS0hDYmZzWkxtdHBvT09QckwKTFE1ZGZGZTZrQ3AxVUpHZlB5R3dNZWg1aVYwbEdkdExLODg0SHVnUWszSFA1a2VVZHhnU0FoVzRJcy9OVHFuLwpybjhQcXY4ZkRWTDdJOG1RRXoxYVBSK05KQkVYdEpsS0pwU3hwS2QrVWx3a1hmcGtnZDg9Ci0tLS0tRU5EIFJTQSBQUklWQVRFIEtFWS0tLS0tCg==
-kind: Secret
-metadata:
-  creationTimestamp: "2024-01-08T17:25:39Z"
-  name: user-token-signing-key-1
-  namespace: kuma-system
-type: system.kuma.io/global-secret
