--- app/kumactl/cmd/export/testdata/export.golden.yaml
@@ -1,30 +1,0 @@
----
-creationTime: "0001-01-01T00:00:00Z"
-modificationTime: "0001-01-01T00:00:00Z"
-name: default
-type: Mesh
----
-creationTime: "0001-01-01T00:00:00Z"
-modificationTime: "0001-01-01T00:00:00Z"
-name: another-mesh
-type: Mesh
----
-creationTime: "0001-01-01T00:00:00Z"
-modificationTime: "0001-01-01T00:00:00Z"
-name: user-token-signing-key-1
-type: GlobalSecret
-data: LS0tLS1CRUdJTiBSU0EgUFJJVkFURSBLRVktLS0tLQpNSUlFb3dJQkFBS0NBUUVBMmhWV3F5MUp6cGJNRUlyVHIyVVRkRXNDTXJUYVVoQXFabGhOTEpyUzhrYzBtYlBLClo3blQxSkVySkdVSTQvTlJIU1FRWk5EM2JpUEVjVmt0dTBQNVVpNklldW04SEg3RU5vUmFTd0xnYVY0NTFMT0cKNWNXeU9ickt5Q0FlaDFpcHJtWDZZclAwVTlOQnZMN0YrUm83eGpoVXAvR0NzRVBadC9VQjR6WGd6ZjNYeHhWUwpmZ1hLMHB0RzNyQ3AyTjlTcEIvdVo0TWhEaGROLys0ZkRkMWhsaEpXRWpCS0t2T0h4M2ZPcHFiWEFCbEhrTDNJCmhBUWJtYlYzZlVqVFcyamZPNGJvaU8rK0lUQjc0TUdyNTU2Rm95aEJpTEZwQzd2OGFEdldQS0ZybGM3SVZOMkYKOHNVaisxK2l3QTBnMnA0YnRRcWJwQTZDcHpXSitCUzdBemNnclFJREFRQUJBb0lCQUVKdCs0OTY1UitieVo5RQpXb1R6MmdBemxRRFRIQ21oS1BreDdwZ210Umo5M2FUd0lqaUVYMlZCcncyUUJTeENnclBJdDRHdU1hQ0lIcCtVClZaelVjTTY2dy9JRTNyc2pWaXR4RjN2LzdWSExWUHhZanovUWY2Q2NLSWJiV2RGS2phMUUvUWN4Mzd2ay93STYKYmwyWnd4bDNNVHBURThjVGMrOVlIM0lSaUYvV1lkODhwZVJoRWc2QjBrMmVyU2NKWXp1Vm9qRC9LVHBhWWl2OAp2WG5BZWttMlNUWDdJSnBwd0JLZkRjYWZ4V0FDRzFNSEZFeDI4NFpzKzBhdkdZc1hZTWh0bUQ3bEpPRTVpR2ZFClVnQzNhc3N0UDh3eGRCYzVNNEZCd1dUOXpveEhwa09hSXNORFMxWlkzMlJ0ZXB1eUF0cUljb1BMOTNvTFkvQUkKQkNQekd4VUNnWUVBOW45WkpyRGE4ZGRabXY4amxKNm1ldVo2T1doWjlWcktNYjRuZXZTc29EekdyTDBQK0wrTgo4T2Zqb0JLVk9NMVRCSW1QSXBCbGE2d25xMDQ1ZGtmeGkwVTFWYkpVS00xU0Y3cHVHSTAwMnJhVjN4Tks4MStLCkRSS0xoVGtNN21uaHRYREpNTGZrVkE4czM0YlYrc01GSUpDdU13dzBhUlpqMUl1U2xUNHB3dXNDZ1lFQTRuMlQKUDdxcHlMYXRRODJWMkMxbmVxcWs3WG9tRFRtQnA3K25uVTd2YmtMdnVxcXgwa1B2QngxWlhnUjVmWG1iUmI2MQp6bDF6YTR3SkttZHRwQU94cHVJRGk2QUdaUjVXVXN1ZGJpMCtBbGxBbUdCdUl5NklHZG94WkVkSlppMXpGbW9MCk1xQmdrVGdmOTJMMVV2eGRhR1NudjFFZmVRTTFYL2xXMEdDQzFNY0NnWUJPMkk5SXUvdGNyamlnUm9vNENNTlIKdnBWTDlhQUl4a0MweHBGVVdCMDdxQlVZdmVUMCtUakszVEhIU3dwbVArOFZIWGNtWkMrODBEVmtLMEpnSGtVWgora3BFSGlXalk0R2x5MkZhQXhqcitNbFpzSThoNStaOXBmZnc1ajNIcVNDa2ErNCs0aU5yNitxN21TS3E4bVNhCmRRenprU3pXakcza1Y3ek1kalRmZndLQmdRREVZQ3JtNjdyQVc1dDVHV3lkeGloZGxFMXgxaUhkLzRscVgrTTkKb3E0NGhQTUVmM0cwcE5BLzgvT0RjQS9DZWtqbHZRZzNpSkZGbzJKakVOQmdrQTlkYWNPM01TVzZVUVo1dlMrZgowcnhjSHZUQjZZSTFJT2hDdWd0dkxkdjNLdHNKaVNKeVg5bTZyOVlQa2c2eFNEelpBY1RzUGJNeklsUUtCYzkrCjF0ODVDd0tCZ0dFeWlOSG5NRHNwUVlBTk5UazdvY0syRXlCM2lTL2ZkcjVXcXg1anpucWRCZm0wTmZZRE5FcHcKemswQ1doUm10STdITVNuRlRRT3o5NDIxTWlNTzhDbkxaTHF4UkJ4cm8vSHI1a2I3RnFtSTQvSXRVUktiSlJlSwo2a0tZT2JNbCsyU0w2MUQ2MjVCNnVBSDR0WEhpUmEyWGQ1ZWZyWm9nRmpiSkUwNnpFb3lKCi0tLS0tRU5EIFJTQSBQUklWQVRFIEtFWS0tLS0t
----
-creationTime: "0001-01-01T00:00:00Z"
-mesh: default
-modificationTime: "0001-01-01T00:00:00Z"
-name: dataplane-token-signing-key-1
-type: Secret
-data: LS0tLS1CRUdJTiBSU0EgUFJJVkFURSBLRVktLS0tLQpNSUlFb3dJQkFBS0NBUUVBMmhWV3F5MUp6cGJNRUlyVHIyVVRkRXNDTXJUYVVoQXFabGhOTEpyUzhrYzBtYlBLClo3blQxSkVySkdVSTQvTlJIU1FRWk5EM2JpUEVjVmt0dTBQNVVpNklldW04SEg3RU5vUmFTd0xnYVY0NTFMT0cKNWNXeU9ickt5Q0FlaDFpcHJtWDZZclAwVTlOQnZMN0YrUm83eGpoVXAvR0NzRVBadC9VQjR6WGd6ZjNYeHhWUwpmZ1hLMHB0RzNyQ3AyTjlTcEIvdVo0TWhEaGROLys0ZkRkMWhsaEpXRWpCS0t2T0h4M2ZPcHFiWEFCbEhrTDNJCmhBUWJtYlYzZlVqVFcyamZPNGJvaU8rK0lUQjc0TUdyNTU2Rm95aEJpTEZwQzd2OGFEdldQS0ZybGM3SVZOMkYKOHNVaisxK2l3QTBnMnA0YnRRcWJwQTZDcHpXSitCUzdBemNnclFJREFRQUJBb0lCQUVKdCs0OTY1UitieVo5RQpXb1R6MmdBemxRRFRIQ21oS1BreDdwZ210Umo5M2FUd0lqaUVYMlZCcncyUUJTeENnclBJdDRHdU1hQ0lIcCtVClZaelVjTTY2dy9JRTNyc2pWaXR4RjN2LzdWSExWUHhZanovUWY2Q2NLSWJiV2RGS2phMUUvUWN4Mzd2ay93STYKYmwyWnd4bDNNVHBURThjVGMrOVlIM0lSaUYvV1lkODhwZVJoRWc2QjBrMmVyU2NKWXp1Vm9qRC9LVHBhWWl2OAp2WG5BZWttMlNUWDdJSnBwd0JLZkRjYWZ4V0FDRzFNSEZFeDI4NFpzKzBhdkdZc1hZTWh0bUQ3bEpPRTVpR2ZFClVnQzNhc3N0UDh3eGRCYzVNNEZCd1dUOXpveEhwa09hSXNORFMxWlkzMlJ0ZXB1eUF0cUljb1BMOTNvTFkvQUkKQkNQekd4VUNnWUVBOW45WkpyRGE4ZGRabXY4amxKNm1ldVo2T1doWjlWcktNYjRuZXZTc29EekdyTDBQK0wrTgo4T2Zqb0JLVk9NMVRCSW1QSXBCbGE2d25xMDQ1ZGtmeGkwVTFWYkpVS00xU0Y3cHVHSTAwMnJhVjN4Tks4MStLCkRSS0xoVGtNN21uaHRYREpNTGZrVkE4czM0YlYrc01GSUpDdU13dzBhUlpqMUl1U2xUNHB3dXNDZ1lFQTRuMlQKUDdxcHlMYXRRODJWMkMxbmVxcWs3WG9tRFRtQnA3K25uVTd2YmtMdnVxcXgwa1B2QngxWlhnUjVmWG1iUmI2MQp6bDF6YTR3SkttZHRwQU94cHVJRGk2QUdaUjVXVXN1ZGJpMCtBbGxBbUdCdUl5NklHZG94WkVkSlppMXpGbW9MCk1xQmdrVGdmOTJMMVV2eGRhR1NudjFFZmVRTTFYL2xXMEdDQzFNY0NnWUJPMkk5SXUvdGNyamlnUm9vNENNTlIKdnBWTDlhQUl4a0MweHBGVVdCMDdxQlVZdmVUMCtUakszVEhIU3dwbVArOFZIWGNtWkMrODBEVmtLMEpnSGtVWgora3BFSGlXalk0R2x5MkZhQXhqcitNbFpzSThoNStaOXBmZnc1ajNIcVNDa2ErNCs0aU5yNitxN21TS3E4bVNhCmRRenprU3pXakcza1Y3ek1kalRmZndLQmdRREVZQ3JtNjdyQVc1dDVHV3lkeGloZGxFMXgxaUhkLzRscVgrTTkKb3E0NGhQTUVmM0cwcE5BLzgvT0RjQS9DZWtqbHZRZzNpSkZGbzJKakVOQmdrQTlkYWNPM01TVzZVUVo1dlMrZgowcnhjSHZUQjZZSTFJT2hDdWd0dkxkdjNLdHNKaVNKeVg5bTZyOVlQa2c2eFNEelpBY1RzUGJNeklsUUtCYzkrCjF0ODVDd0tCZ0dFeWlOSG5NRHNwUVlBTk5UazdvY0syRXlCM2lTL2ZkcjVXcXg1anpucWRCZm0wTmZZRE5FcHcKemswQ1doUm10STdITVNuRlRRT3o5NDIxTWlNTzhDbkxaTHF4UkJ4cm8vSHI1a2I3RnFtSTQvSXRVUktiSlJlSwo2a0tZT2JNbCsyU0w2MUQ2MjVCNnVBSDR0WEhpUmEyWGQ1ZWZyWm9nRmpiSkUwNnpFb3lKCi0tLS0tRU5EIFJTQSBQUklWQVRFIEtFWS0tLS0t
----
-creationTime: "0001-01-01T00:00:00Z"
-mesh: another-mesh
-modificationTime: "0001-01-01T00:00:00Z"
-name: dataplane-token-signing-key-1
-type: Secret
-data: LS0tLS1CRUdJTiBSU0EgUFJJVkFURSBLRVktLS0tLQpNSUlFb3dJQkFBS0NBUUVBMmhWV3F5MUp6cGJNRUlyVHIyVVRkRXNDTXJUYVVoQXFabGhOTEpyUzhrYzBtYlBLClo3blQxSkVySkdVSTQvTlJIU1FRWk5EM2JpUEVjVmt0dTBQNVVpNklldW04SEg3RU5vUmFTd0xnYVY0NTFMT0cKNWNXeU9ickt5Q0FlaDFpcHJtWDZZclAwVTlOQnZMN0YrUm83eGpoVXAvR0NzRVBadC9VQjR6WGd6ZjNYeHhWUwpmZ1hLMHB0RzNyQ3AyTjlTcEIvdVo0TWhEaGROLys0ZkRkMWhsaEpXRWpCS0t2T0h4M2ZPcHFiWEFCbEhrTDNJCmhBUWJtYlYzZlVqVFcyamZPNGJvaU8rK0lUQjc0TUdyNTU2Rm95aEJpTEZwQzd2OGFEdldQS0ZybGM3SVZOMkYKOHNVaisxK2l3QTBnMnA0YnRRcWJwQTZDcHpXSitCUzdBemNnclFJREFRQUJBb0lCQUVKdCs0OTY1UitieVo5RQpXb1R6MmdBemxRRFRIQ21oS1BreDdwZ210Umo5M2FUd0lqaUVYMlZCcncyUUJTeENnclBJdDRHdU1hQ0lIcCtVClZaelVjTTY2dy9JRTNyc2pWaXR4RjN2LzdWSExWUHhZanovUWY2Q2NLSWJiV2RGS2phMUUvUWN4Mzd2ay93STYKYmwyWnd4bDNNVHBURThjVGMrOVlIM0lSaUYvV1lkODhwZVJoRWc2QjBrMmVyU2NKWXp1Vm9qRC9LVHBhWWl2OAp2WG5BZWttMlNUWDdJSnBwd0JLZkRjYWZ4V0FDRzFNSEZFeDI4NFpzKzBhdkdZc1hZTWh0bUQ3bEpPRTVpR2ZFClVnQzNhc3N0UDh3eGRCYzVNNEZCd1dUOXpveEhwa09hSXNORFMxWlkzMlJ0ZXB1eUF0cUljb1BMOTNvTFkvQUkKQkNQekd4VUNnWUVBOW45WkpyRGE4ZGRabXY4amxKNm1ldVo2T1doWjlWcktNYjRuZXZTc29EekdyTDBQK0wrTgo4T2Zqb0JLVk9NMVRCSW1QSXBCbGE2d25xMDQ1ZGtmeGkwVTFWYkpVS00xU0Y3cHVHSTAwMnJhVjN4Tks4MStLCkRSS0xoVGtNN21uaHRYREpNTGZrVkE4czM0YlYrc01GSUpDdU13dzBhUlpqMUl1U2xUNHB3dXNDZ1lFQTRuMlQKUDdxcHlMYXRRODJWMkMxbmVxcWs3WG9tRFRtQnA3K25uVTd2YmtMdnVxcXgwa1B2QngxWlhnUjVmWG1iUmI2MQp6bDF6YTR3SkttZHRwQU94cHVJRGk2QUdaUjVXVXN1ZGJpMCtBbGxBbUdCdUl5NklHZG94WkVkSlppMXpGbW9MCk1xQmdrVGdmOTJMMVV2eGRhR1NudjFFZmVRTTFYL2xXMEdDQzFNY0NnWUJPMkk5SXUvdGNyamlnUm9vNENNTlIKdnBWTDlhQUl4a0MweHBGVVdCMDdxQlVZdmVUMCtUakszVEhIU3dwbVArOFZIWGNtWkMrODBEVmtLMEpnSGtVWgora3BFSGlXalk0R2x5MkZhQXhqcitNbFpzSThoNStaOXBmZnc1ajNIcVNDa2ErNCs0aU5yNitxN21TS3E4bVNhCmRRenprU3pXakcza1Y3ek1kalRmZndLQmdRREVZQ3JtNjdyQVc1dDVHV3lkeGloZGxFMXgxaUhkLzRscVgrTTkKb3E0NGhQTUVmM0cwcE5BLzgvT0RjQS9DZWtqbHZRZzNpSkZGbzJKakVOQmdrQTlkYWNPM01TVzZVUVo1dlMrZgowcnhjSHZUQjZZSTFJT2hDdWd0dkxkdjNLdHNKaVNKeVg5bTZyOVlQa2c2eFNEelpBY1RzUGJNeklsUUtCYzkrCjF0ODVDd0tCZ0dFeWlOSG5NRHNwUVlBTk5UazdvY0syRXlCM2lTL2ZkcjVXcXg1anpucWRCZm0wTmZZRE5FcHcKemswQ1doUm10STdITVNuRlRRT3o5NDIxTWlNTzhDbkxaTHF4UkJ4cm8vSHI1a2I3RnFtSTQvSXRVUktiSlJlSwo2a0tZT2JNbCsyU0w2MUQ2MjVCNnVBSDR0WEhpUmEyWGQ1ZWZyWm9nRmpiSkUwNnpFb3lKCi0tLS0tRU5EIFJTQSBQUklWQVRFIEtFWS0tLS0t
