+++ pkg/api-server/policies_endpoints.go
@@ -1,13 +1,11 @@
 package api_server
 
 import (
-	"github.com/kumahq/kuma/pkg/core/resources/apis/mesh"
 	"sort"
 
 	"github.com/emicklei/go-restful/v3"
 
-	api_types "github.com/kumahq/kuma/api/openapi/types"
-	api_common "github.com/kumahq/kuma/api/openapi/types/common"
+	"github.com/kumahq/kuma/pkg/api-server/mappers"
 	"github.com/kumahq/kuma/pkg/api-server/types"
 	"github.com/kumahq/kuma/pkg/core/resources/model"
 )
@@ -40,38 +38,9 @@ func addPoliciesWsEndpoints(ws *restful.WebService, federatedZone bool, readOnly
 		}
 	}))
 	ws.Route(ws.GET("/_resources").To(func(req *restful.Request, resp *restful.Response) {
-		response := api_types.ResourceTypeDescriptionList{}
-		for _, def := range defs {
-			td := api_common.ResourceTypeDescription{
-				Name:                            string(def.Name),
-				ReadOnly:                        readOnly || federatedZone || def.ReadOnly,
-				Path:                            def.WsPath,
-				SingularDisplayName:             def.SingularDisplayName,
-				PluralDisplayName:               def.PluralDisplayName,
-				Scope:                           api_common.ResourceTypeDescriptionScope(def.Scope),
-				IncludeInFederation: includeInFederation(def.KDSFlags, def.Name),
-			}
-			if def.IsPolicy {
-				td.Policy = &api_common.PolicyDescription{
-					HasToTargetRef:   def.HasToTargetRef,
-					HasFromTargetRef: def.HasFromTargetRef,
-					IsTargetRef:      def.IsTargetRefBased,
-				}
-			}
-			response.Resources = append(response.Resources, td)
-		}
-		sort.SliceStable(response.Resources, func(i, j int) bool {
-			return response.Resources[i].Name < response.Resources[j].Name
-		})
+		response := mappers.MapResourceTypeDescription(defs, readOnly, federatedZone)
 		if err := resp.WriteAsJson(response); err != nil {
 			log.Error(err, "Could not write the response")
 		}
 	}))
-}
-
-func includeInFederation(kdsFlags model.KDSFlagType, name model.ResourceType) bool {
-	if name == mesh.ServiceInsightType {
-		return false
-	}
-	return kdsFlags & model.GlobalToAllZonesFlag != 0
 }
